+++ profile/profiledef.sh
@@ -3,8 +3,8 @@
 
 iso_name="warble-linux"
 iso_label="WARBLE_$(date +%Y%m)"
-iso_publisher="Warble Linux <https://github.com/warble-linux>"
-iso_application="Warble-Linux Sandbox OS Live/Rescue CD"
+iso_publisher="Warble Linux <https://github.com/warble-tech/warble-linux>"
+iso_application="Warble Linux live sandbox"
 iso_version="$(date +%Y.%m.%d)"
 install_dir="arch"
 buildmodes=('iso')